Update 20 Dec 2020
Added additional tracking and color changes to pickaxes.

Added additional tracking for swords.

Added tracking for bows.

Tools now require StatTrak to be applied in order to function. StatTrak tools by renaming a name tag to "StatTrak" in an anvil, and dropping the modified name tag on top of an anvil with your tool of choice and 5 levels.